+++ src/components/searchResult/Transaction.InnerTransactions.vue
@@ -111,6 +111,18 @@ export default {
             info.details.push({ key: 'Message', value: this.params[index].message.payload })
           }
 
+          let amountXpx
+          if (this.params[index].mosaics && this.params[index].mosaics.length > 0) {
+            amountXpx = this.params[index].mosaics.filter(el => el.id.toHex() === this.$store.state.xpx)
+            if (amountXpx.length > 0) {
+              amountXpx = this.$utils.fmtDivisibility(amountXpx[0].amount.compact(), 6)
+              info.details.push({ key: 'Amount', valueHtml: amountXpx })
+            } else {
+              amountXpx = this.$utils.fmtDivisibility(0, 6)
+              info.details.push({ key: 'Amount', valueHtml: amountXpx })
+            }
+          }
+
           break;
 
         case 'Mosaic definition':